pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spineapi"
version = "0.1.0"
description = "AI-powered backend scaffolding tool that transforms OpenAPI/Swagger specs into production-ready backend applications"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "SpineAPI Team", email = "team@spineapi.dev"}
]
keywords = ["openapi", "swagger", "fastapi", "code-generation", "backend", "scaffolding", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "typer[all]>=0.9.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "openapi-spec-validator>=0.6.0",
    "fastapi>=0.104.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "loguru>=0.7.0",
    "prometheus-client>=0.17.0",
    "redis>=5.0.0",
    "requests>=2.31.0",
    "pydantic>=2.4.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "psycopg2-binary>=2.9.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]

[project.optional-dependencies]
llm = [
    "openai>=1.0.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
]
monitoring = [
    "grafana-api>=1.0.3",
    "prometheus-api-client>=0.5.3",
]
dev = [
    "pre-commit>=3.4.0",
    "tox>=4.11.0",
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/spineapi/spineapi"
Documentation = "https://spineapi.readthedocs.io"
Repository = "https://github.com/spineapi/spineapi.git"
"Bug Tracker" = "https://github.com/spineapi/spineapi/issues"

[project.scripts]
spineapi = "spineapi.cli:app"

[tool.setuptools.packages.find]
include = ["spineapi*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=spineapi --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["spineapi"]
omit = ["*/tests/*", "*/venv/*", "*/env/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
